+++ test/deps.ts
@@ -1,5 +1,6 @@
 export { expect } from "https://deno.land/x/expect@285caf/mod.ts";
-export * as Opine from "https://deno.land/x/opine@0.16.0/mod.ts";
-export * as OpineTypes from "https://deno.land/x/opine@0.16.0/src/types.ts";
+export * as Opine from "https://deno.land/x/opine@0.17.0/mod.ts";
+export * as OpineTypes from "https://deno.land/x/opine@0.17.0/src/types.ts";
 export * as Oak from "https://deno.land/x/oak@v5.3.1/mod.ts";
-export { dirname, join } from "https://deno.land/std@0.59.0/path/mod.ts";
+export * as Ako from "https://deno.land/x/ako@5b5980/mod.ts";
+export { dirname, join } from "https://deno.land/std@0.60.0/path/mod.ts";